--- a/packages/babel-preset-gatsby/src/__tests__/index.js
+++ b/packages/babel-preset-gatsby/src/__tests__/index.js
@@ -8,6 +8,7 @@ it(`Specifies proper presets and plugins for test stage`, () => {
     [
       expect.stringContaining(path.join(`@babel`, `preset-env`)),
       {
+        exclude: [`transform-typeof-symbol`],
         corejs: 2,
         loose: true,
         modules: `commonjs`,
@@ -40,8 +41,13 @@ it(`Specifies proper presets and plugins for test stage`, () => {
     [
       expect.stringContaining(path.join(`@babel`, `plugin-transform-runtime`)),
       {
+        absoluteRuntimePath: expect.stringContaining(
+          path.join(`@babel`, `runtime`)
+        ),
+        corejs: false,
         helpers: true,
         regenerator: true,
+        useESModules: false,
       },
     ],
   ])
@@ -63,6 +69,7 @@ it(`Specifies proper presets and plugins for build-html stage`, () => {
     [
       expect.stringContaining(path.join(`@babel`, `preset-env`)),
       {
+        exclude: [`transform-typeof-symbol`],
         corejs: 2,
         loose: true,
         modules: false,
@@ -95,8 +102,13 @@ it(`Specifies proper presets and plugins for build-html stage`, () => {
     [
       expect.stringContaining(path.join(`@babel`, `plugin-transform-runtime`)),
       {
-        helpers: true,
+        absoluteRuntimePath: expect.stringContaining(
+          path.join(`@babel`, `runtime`)
+        ),
+        helpers: false,
         regenerator: true,
+        corejs: false,
+        useESModules: true,
       },
     ],
   ])
@@ -111,6 +123,7 @@ it(`Allows to configure browser targets`, () => {
   expect(presets[0]).toEqual([
     expect.stringContaining(path.join(`@babel`, `preset-env`)),
     {
+      exclude: [`transform-typeof-symbol`],
       corejs: 2,
       loose: true,
       modules: false,