--- a/packages/babel-preset-gatsby/src/dependencies.js
+++ b/packages/babel-preset-gatsby/src/dependencies.js
@@ -25,6 +25,8 @@ module.exports = function(api, options = {}) {
           useBuiltIns: `usage`,
           corejs: 2,
           modules: false,
+          // Exclude transforms that make all code slower (https://github.com/facebook/create-react-app/pull/5278)
+          exclude: [`transform-typeof-symbol`],
         },
       ],
     ].filter(Boolean),