--- a/packages/gatsby/src/utils/webpack.config.js
+++ b/packages/gatsby/src/utils/webpack.config.js
@@ -248,12 +248,6 @@ module.exports = async (program, directory, suppliedStage) => {
   function getModule() {
     const jsOptions = {}
 
-    // Speedup 🏎️💨 the build! We only include transpilation of node_modules on javascript production builds
-    // TODO create gatsby plugin to enable this behaviour on develop (only when people are requesting this feature)
-    if (stage !== `build-javacript`) {
-      jsOptions.exclude = [`node_modules`]
-    }
-
     // Common config for every env.
     // prettier-ignore
     let configRules = [
@@ -264,6 +258,13 @@ module.exports = async (program, directory, suppliedStage) => {
       rules.media(),
       rules.miscAssets(),
     ]
+
+    // Speedup 🏎️💨 the build! We only include transpilation of node_modules on javascript production builds
+    // TODO create gatsby plugin to enable this behaviour on develop (only when people are requesting this feature)
+    if (stage === `build-javascript`) {
+      configRules.push(rules.dependencies())
+    }
+
     if (store.getState().themes.themes) {
       configRules = configRules.concat(
         store.getState().themes.themes.map(theme => {