… packages/gatsby/src/utils/__tests__/__snapshots__/webpack-utils.js.snap
… yarn.lock