## Description
Implements point 1 & 2 from #15213. Mostly stolen from create-react-app.

We revert the changes done by https://github.com/gatsbyjs/gatsby/pull/14111 in the js loader and create a new loader called `dependencies` to do specific work on node_modules.

Inside the `dependencies` loader, we set every optimisation to false. We also disable sourcemap generation as this might fail on big files (ex #15190).

It's not really much faster.

| site | before node transpilation | after node transpilation | This PR |
|-----------------|---------------------------|--------------------------|---------|
| gatsbyjs.org | 88.907s | 110s | 91.936s |
| moonmeister.net | 55.749s | 82s | 57.216s |
| default-starter | 9.444s | 13s | 13s |

More PRs will follow to speedup webpack build time. Sadly most of these PRs will only work on next build.

## Related Issues
Fixes #15190
Fixes #15183
Fixes #15207